Add Start Menu
This commit adds a start menu where the player can choose between
a 1- or 2-player game, as well as a button in game to return to
the menu.

On a personal note. I have been using React for about a week now,
and every time I open my editor I am amazed (again) at how simple
it is to implement the ideas I have in my head.
Major kudos to the React
Team.