Add AI/Finish project
This update functionally finishes the project. There are still
some rough bits to polish up: When starting a match against
the Ai, with the AI taking the first move, it's really slow to
load. I think that's because I'm running the minimax function
in ComponentDidMount, but I need to do some more reading and
tinkering to be sure.